+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 
 ## patch
 * allow seo meta keywords and hashtags to be selected using comma key
+* style (in triniti) the save buttons
 * Add fix for cutting and pasting a text block
 * style (in triniti) the key values field select component
+* Add divider block into blocksmith
 * style video captions selector
 * update `changedDate` and `changedTime` tests to not use `new Date()` because then it may fail if you do it at exactly the right (wrong) time.
+* style fragments selector
 
 
 ## v0.1.1